Charts: changed input and labels to arrays

+++ src/MudBlazor/Components/Chart/Models/ChartSeries.cs
@@ -8,6 +8,6 @@ namespace MudBlazor
     {
         public string Name { get; set; }
 
-        public int[] Data { get; set; }
+        public double[] Data { get; set; }
     }
 }
